motor fix

+++ src/firmware/control/angles/angles.c
@@ -12,7 +12,7 @@
 #include <math.h>
 
 void imu2euler(double* ypr, double accX, double accY, double accZ, double magX, double magY){
-        ypr[YAW_ANGLE] = (atan2f(magY,magX)*180.0)/M_PI;
+        ypr[YAW_ANGLE] = (atan2(magY,magX)*180.0)/M_PI;
         ypr[PITCH_ANGLE] = (atan2(accX, sqrt(accY*accY + accZ*accZ))*180.0)/M_PI;
         ypr[ROLL_ANGLE] = (atan2(-accY, accZ)*180.0)/M_PI;
 }
