+++ src/firmware/protocols/accel/accel.h
@@ -1,8 +1,8 @@
-/* 
-	PAVx -- Pod-Based Autonomous Vehicles 
+/*
+	PAVx -- Pod-Based Autonomous Vehicles
 	Library Created By: Sargis S Yonan
 	March 2017
-*/ 
+*/
 
 #ifndef _ACCEL_H_
 #define _ACCEL_H_
@@ -13,10 +13,10 @@ void InitializeAccel(void);
 
 void Accel_Update(void);
 
-float Accel_GetX(void);
+double Accel_GetX(void);
 
-float Accel_GetY(void);
+double Accel_GetY(void);
 
-float Accel_GetZ(void);
+double Accel_GetZ(void);
 
 #endif
